+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - FieldSets are no longer defined in a submodule `field_sets`. They're now put in the same root as the reset of the generated code.
 - Removed the 'read_all_registers' functions and the accompanying enums
 - Updated `convert_case` to 0.8
+- Updated `defmt` to 1.0.1. The crate feature is now called `defmt` instead of `defmt-03`
 
 ### 1.0.7 (30-07-25)